+++ packages/ironbean/src/config.ts
@@ -1,0 +1,84 @@
+import {take, Container, ApplicationContext, DependencyToken} from "./internals";
+
+type MapToType = {
+    string: string;
+    number: number;
+    boolean: boolean;
+}
+
+type PrimitiveKeys = keyof MapToType;
+
+interface ConfigParams {
+    [key: string]: PrimitiveKeys | ConfigParams
+}
+
+type Schema<T> = {
+    [P in keyof T]: T[P] extends PrimitiveKeys ? MapToType[T[P]] : Schema<T[P]>
+}
+
+type Config<T> = ConfigChild<T> & {
+    apply(context: ApplicationContext, values: Schema<T>): void
+}
+
+type ConfigChild<T> = {
+    [P in keyof T]: T[P] extends "string" ? DependencyToken<string> :
+                    T[P] extends "number" ? DependencyToken<number> :
+                    T[P] extends "boolean" ? DependencyToken<boolean> :
+                    T[P] extends {} ? ConfigChild<T[P]> :
+                    never;
+};
+
+export function createConfig<T extends ConfigParams>(params: T): Config<T> {
+    const cfg = _createConfig(params);
+    Object.defineProperty(cfg, "apply", {
+        value: (context: ApplicationContext, values: any) => {
+            const container = context.getBean(Container);
+            fillConfig(container, cfg, values);
+        },
+        enumerable: false,
+    });
+
+    return cfg;
+}
+
+function _createConfig(params: any, path: string = ""): any {
+    const config: any = {};
+
+    for (let key in params) {
+        const fullPath = path ? `${path}.${key}` : key;
+        config[key] = resolveValue(params[key], fullPath);
+    }
+
+    return config;
+}
+
+function resolveValue(value: PrimitiveKeys | object, name: string): any {
+    if (typeof value === "string") {
+        const token = DependencyToken.create(name);
+        switch (value) {
+            case "string":
+                take(token).setClassType(String);
+                break;
+            case "number":
+                take(token).setClassType(Number);
+                break;
+            case "boolean":
+                take(token).setClassType(Boolean);
+                break;
+        }
+
+        return token;
+    }
+    return _createConfig(value, name);
+}
+
+function fillConfig(container: Container, config: any, values: any) {
+    for (let key in config) {
+        const value = config[key];
+        if (value instanceof DependencyToken) {
+            container.setValue(value as any, values[key]);
+        } else {
+            fillConfig(container, value as any, values[key])
+        }
+    }
+}